backend-ia/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
langchain==0.3.14
langchain-community==0.3.14
langchain-experimental==0.3.4
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
httpx==0.28.1
python-dotenv==1.0.1
